pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image_ranking_pipeline"
version = "0.1.0"
description = "Minimal image ranking pipeline scaffolding with CLI for metadata extraction"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
  "Pillow>=10.0.0",
  "piexif>=1.1.3",
  "exifread>=3.0.0",
  "pillow-heif>=0.16.0; platform_system != 'Windows' or platform_machine != 'ARM64'",
  "click>=8.1.7",
  "tqdm>=4.66.0",
  "opencv-python>=4.8.0",
  "ImageHash>=4.3.1",
  "scikit-image>=0.23.0",
  "scikit-learn>=1.3.0",
  "pytest>=7.4.0",
  "numpy>=1.24,<2",
  "torch==2.2.2",
  "torchvision==0.17.2",
  "open_clip_torch>=2.23.0"
]

[project.optional-dependencies]
clip = [
  "open-clip-torch>=2.24.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "scikit-learn>=1.2.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.black]
line-length = 100
target-version = ["py39"]
exclude = "^/(.venv|venv|build|dist|.git)/"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "UP", "B"]
ignore = []
exclude = [".venv", "venv", "build", "dist", ".git"]



